shipping/requirements.txt
alembic==1.4.3
bcrypt==3.2.0
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.3
click==7.1.2
cssselect2==0.3.0
defusedxml==0.6.0
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WeasyPrint==0.6
Flask-WTF==0.14.3
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
jellyfish==0.6.1
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient==2.0.1
Pillow==7.2.0
pycparser==2.20
Pyphen==0.9.5
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
tinycss2==1.0.2
us==2.0.2
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3